requirements.txt
# Backend Python dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sentence-transformers==2.3.1
transformers==4.37.0
torch==2.1.2
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
PyPDF2==3.0.1
python-docx==1.1.0
